# swagger.yaml
openapi: 3.0.0
info:
  title: Talent API
  version: 1.0.0
  description: >
    REST API til Tech Chapter praktikant-udfordring.
    Returnerer talent-profiler og tilhørende dokumenter (CV, motivationsbrev, projektbeskrivelser).
    Bygget med Spring Boot 3, Java 21, Lombok og Swagger/OpenAPI 3.
  contact:
    name: Andreas Gabel
    url: https://andreasgabel.dk
    email: andreasgabel98@gmail.com
servers:
  - url: /
    description: Current server
paths:
  /talent:
    get:
      summary: Hent alle talents
      description: Returnerer en liste over alle registrerede talent-profiler med kontaktinfo og links.
      operationId: getAllTalents
      tags:
        - Talent
      responses:
        '200':
          description: Liste af talents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Talent'
  /talent/{id}:
    get:
      summary: Hent en specifik talent
      description: Returnerer en enkelt talent-profil ud fra det angivne ID (UUID).
      operationId: getTalentById
      tags:
        - Talent
      parameters:
        - name: id
          in: path
          required: true
          description: Talent ID (UUID)
          example: 550e8400-e29b-41d4-a716-446655440000
          schema:
            type: string
      responses:
        '200':
          description: Talent fundet
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Talent'
        '404':
          description: Talent ikke fundet — ugyldigt eller ukendt ID
  /talent/{id}/documents:
    get:
      summary: Hent dokumenter for en talent
      description: Returnerer alle dokumenter (CV, motivationsbrev, projektbeskrivelser) der tilhører den angivne talent.
      operationId: getDocuments
      tags:
        - Talent
      parameters:
        - name: id
          in: path
          required: true
          description: Talent ID (UUID)
          example: 550e8400-e29b-41d4-a716-446655440000
          schema:
            type: string
      responses:
        '200':
          description: Liste af dokumenter
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Document'
        '404':
          description: Talent ikke fundet
  /talent/{id}/documents/{documentId}:
    get:
      summary: Hent et specifikt dokument
      description: Returnerer et enkelt dokument for en talent, f.eks. CV eller motivationsbrev.
      operationId: getDocument
      tags:
        - Talent
      parameters:
        - name: id
          in: path
          required: true
          description: Talent ID (UUID)
          example: 550e8400-e29b-41d4-a716-446655440000
          schema:
            type: string
        - name: documentId
          in: path
          required: true
          description: Document ID
          example: d001
          schema:
            type: string
      responses:
        '200':
          description: Dokument fundet
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Document'
        '404':
          description: Talent eller dokument ikke fundet
components:
  schemas:
    Talent:
      type: object
      description: En talent-profil med kontaktoplysninger og links til online profiler
      properties:
        id:
          type: string
          format: uuid
          description: Unik identifikator (UUID)
          example: 550e8400-e29b-41d4-a716-446655440000
        name:
          type: string
          description: Fulde navn
          example: Andreas Gabel
        title:
          type: string
          description: Stillingsbetegnelse eller rolle
          example: Datamatiker-studerende | DevOps & Backend
        profile_text:
          type: string
          description: Kort introduktion til talentet
          example: 4. semester datamatiker med fokus på DevOps og backend...
        email:
          type: string
          format: email
          description: E-mailadresse
          example: andreasgabel98@gmail.com
        phone:
          type: string
          description: Telefonnummer
          example: "+45 12 34 56 78"
        city:
          type: string
          description: By
          example: København
        country:
          type: string
          description: Land
          example: Danmark
        github:
          type: string
          format: uri
          nullable: true
          description: GitHub-profil URL
          example: https://github.com/Gabel1998
        linkedin:
          type: string
          format: uri
          nullable: true
          description: LinkedIn-profil URL
          example: https://www.linkedin.com/in/andreas-søgaard-gabel-758991133
    Document:
      type: object
      description: Et dokument tilknyttet en talent — f.eks. CV, motivationsbrev eller projektbeskrivelse
      properties:
        id:
          type: string
          description: Unik identifikator
          example: d001
        name:
          type: string
          description: Dokumentnavn
          example: Motivationsbrev
        content:
          type: string
          description: Dokumentets indhold i fuld tekst
          example: Kære Tech Chapter, jeg søger hermed...